changes and function
only change in index is the header (h1) into my name. next is to make the funciton writeTweet to have six varibles named (intro, body, end, randIndex, and sentence). for intro place 4 quotes, body 5, and end 2. then in var randIndex make it randomly pick from intro subtracting one because the subtract one is so it won't have a unidentified string because 0 is included. do the same process for body and end. after that make sentence equal to the intro[randIndex];. but for body and end it is writen as (sentence +" "+ body[randIndex];. finally to update it you will have to use document.getElementById("tweet").innerHTML=sentence.